+++ lib/component/auth/buttons/apple_sign_in_button.dart
@@ -24,7 +24,7 @@ class AppleSignInButton extends StatelessWidget {
         icon: icon,
         onPressed: () => context
             .pylon<ArcaneAuthProvider>()
-            .signInWithProvider(ArcaneSignInProviderType.apple),
+            .signInWithProvider(context, ArcaneSignInProviderType.apple),
       );
 }